Support multiple calls to the same argument
Arguments like `-output-http-header` and `-output-http-method` can be called
multiple times to insert additional values. This of course conflicts with
the uniquity of hash key names.

So allow an array of values and call the argument once for each value. The
template casts everything for a flat array, including single strings,
because it's DRYer than calling `is_a?(Array)`.