Amend test syntax
Transpec didn't pick these up, but without changing them, we get errors
about the syntax. The new syntax we use here is described at
http://rspec.info/blog/2012/06/rspecs-new-expectation-syntax/

These error because in the Gemfile we specify the use of the latest version
of puppet-rspec, which then dictates the version of RSpec we use. The
quickest way of checking the version is to look inside the Gemfile.lock
which tells us that we're using 2.99.2, which enforces the use of a
different syntax.