Add options to Gor
This commit adds the ability to set a different binary path than the
default, as we found in upgrading that the path was set to
/usr/local/bin/gor rather than /usr/bin/gor so the service failed to
start.

It also allows adding a hash on environment variables which will be
loaded into the upstart file and loaded when the service starts. This is
due to a bug that we ran into in updating documented here:
https://github.com/buger/gor/wiki/Troubleshooting#gor-is-crashing-with-following-stacktrace